requirements.txt
Django==4.1
Flask==2.2.2
nltk==3.7
pandas==1.4.3
pyembed==1.3.3
python-dotenv==0.20.0
tweepy==4.10.0
